Move current graph model artifacts and runtime into Rust core
