+++ rust_core/tests/fixtures/graph_model_v1.json
@@ -1,0 +1,1 @@
+{"format":"rustmc.graph-model","version":1,"definition":{"dimensions":{},"potentials":[["penalty",{"Binary":["Mul",{"Binary":["Pow",{"Param":"alpha"},{"Const":4.0}]},{"Const":-0.1}]}]],"deterministics":[["signal",{"Add":[{"Param":"alpha"},{"MatVec":{"param_name":"beta","data_key":"X"}}]}]],"priors":[{"Normal":{"name":"alpha","mu":{"Const":0.0},"sigma":{"Const":2.0}}},{"VectorNormal":{"name":"beta","n":2,"mu":0.0,"sigma":1.0}},{"HalfNormal":{"name":"sigma","sigma":{"Const":1.0}}}],"likelihoods":[{"family":"Normal","name":"obs","mu_expr":{"Add":[{"Param":"alpha"},{"MatVec":{"param_name":"beta","data_key":"X"}}]},"sigma":{"Param":"sigma"},"observed_key":"y"}]},"schema":{"vectors":[],"observations":[{"key":"y","kind":{"Observation":{"likelihood":"obs"}},"dim":"obs"}],"matrices":[{"key":"X","kind":{"Matrix":{"n_cols":2}},"dim":"obs"},{"key":"X","kind":{"Matrix":{"n_cols":2}},"dim":"obs"}]}}
